+++ NEWS.md
@@ -17,6 +17,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 * `closed_at` column is `POSIXct` [#35](https://github.com/TanguyBarthelemy/IssueTrackeR/issues/35)
+* `get_still_open()` and `generate_age_mat()` return `ts` and `mts` objects  [#41](https://github.com/TanguyBarthelemy/IssueTrackeR/issues/41)
 
 
 ## [1.4.1] - 2026-07-31